fix session validation for payment success

+++ utils/get-and-validate-session.ts
@@ -17,11 +17,24 @@ import { logError, logInfo } from './logger.ts';
 export async function getAndValidateSession(
   c: Context,
   authRepository: AuthRepository,
+  bySessionId?: {
+    sessionId: string | undefined;
+  },
 ): Promise<Session> {
-  const sessionId = getCookie(c, SESSION_ID_COOKIE_NAME);
-  const sessionAccessToken = getCookie(c, SESSION_ACCESS_TOKEN_COOKIE_NAME);
+  let sessionId: string | undefined;
+  let sessionAccessToken: string | undefined;
+  if (bySessionId === undefined) {
+    sessionId = getCookie(c, SESSION_ID_COOKIE_NAME) as SessionId;
+    sessionAccessToken = getCookie(c, SESSION_ACCESS_TOKEN_COOKIE_NAME);
+  } else {
+    sessionId = bySessionId?.sessionId;
+  }
 
-  if (sessionId === undefined || sessionAccessToken === undefined) {
+  if (sessionId === undefined) {
+    throw new HTTPException(STATUS_CODE.Unauthorized);
+  }
+
+  if (bySessionId === undefined && sessionAccessToken === undefined) {
     throw new HTTPException(STATUS_CODE.Unauthorized);
   }
 
@@ -33,39 +46,43 @@ export async function getAndValidateSession(
     throw new HTTPException(STATUS_CODE.Unauthorized);
   }
 
-  if (session.accessToken !== sessionAccessToken) {
-    if (
-      (Date.now() - session.updatedAt.getTime()) >
-        REUSE_DETECTION_THRESHOLD
-    ) {
-      logError(
-        `malicious activity detected: session access token mismatch for reader ${session.readerId}`,
-      );
-      await authRepository.removeSession(session.id, session.readerId);
-      throw new HTTPException(STATUS_CODE.Unauthorized);
+  if (bySessionId === undefined) {
+    if (session.accessToken !== sessionAccessToken) {
+      if (
+        (Date.now() - session.updatedAt.getTime()) >
+          REUSE_DETECTION_THRESHOLD
+      ) {
+        logError(
+          `malicious activity detected: session access token mismatch for reader ${session.readerId}`,
+        );
+        await authRepository.removeSession(session.id, session.readerId);
+        throw new HTTPException(STATUS_CODE.Unauthorized);
+      }
     }
   }
 
-  if ((Date.now() - session.updatedAt.getTime()) > SESSION_ACCESS_TOKEN_TTL) {
-    logInfo(`session access token expired`);
+  if (bySessionId === undefined) {
+    if ((Date.now() - session.updatedAt.getTime()) > SESSION_ACCESS_TOKEN_TTL) {
+      logInfo(`session access token expired`);
 
-    const newAccessToken = generateUUID();
-    const updatedSession: Session = {
-      ...session,
-      accessToken: newAccessToken,
-      updatedAt: new Date(),
-    };
-    session = updatedSession;
+      const newAccessToken = generateUUID();
+      const updatedSession: Session = {
+        ...session,
+        accessToken: newAccessToken,
+        updatedAt: new Date(),
+      };
+      session = updatedSession;
 
-    await authRepository.updateSession(updatedSession);
+      await authRepository.updateSession(updatedSession);
 
-    setCookie(c, SESSION_ACCESS_TOKEN_COOKIE_NAME, newAccessToken, {
-      httpOnly: true,
-      sameSite: 'Lax',
-      secure: true,
-      path: '/',
-      maxAge: SESSION_MAX_AGE_SECONDS,
-    });
+      setCookie(c, SESSION_ACCESS_TOKEN_COOKIE_NAME, newAccessToken, {
+        httpOnly: true,
+        sameSite: 'Lax',
+        secure: true,
+        path: '/',
+        maxAge: SESSION_MAX_AGE_SECONDS,
+      });
+    }
   }
 
   return session;
